Loading screens!
Modified config.cpp to include loading screens on all* maps. MEUPIC*.paa is 3 different pictures(2048x2048 pixels) for use in the maps. Important note however is that not all the maps work because there are currently 64 maps scattered across the modpacks so I dont have all the class names set up but this config covers a good chunk.