Emergency medical kit
Added the emergency medical kit, it has a mass equivalent of 8 biofoam and 1 2.5kl blood bag. The medkit has a new function associated with it that calls the ace bandage function 8 times, has a higher effectiveness for the bandaging aspect, gives a ibuprofen, and pops a 2.5kl blood bag all at once with a 10 second treatment time.