Minor fix to the broken throwable bubbleshield
I managed to fix the throwable bubbleshield to the point where you can throw it and it has a 30sec timer but for some reason when the bubble spawns it doesnt actually block any bullets. I had to swap out some of the class names because it broke the normal bubbeshield if it was left as is.