Probable fix to the drop pod thrusters
Having an issue where drop pod thrusters arent working in server and the rear thruster didnt work, I reworked some of the variables to see if that works. I also reworked the calculations because the pods face south when dropping from the frigate vs north from when i tested them with zeus module.