Broken Visors
Staging the necessary changes for a mass change to the way we handle broken visors. Added a RSCTitle class for the visor itself, can be changed to any paa that is 2048x2048. Added the function/sqf for the actual display of the visor. I also modified my personal broken helmet to be used as a format for how the rest of the broken visors should look. In order for the visors to work the mission builders will have to add the event handler that I also pushed to all the slots. There may need to be additional event handlers for removing the rscresource on death but testing needs to be done first.