Officer Handgun Change / Vehicle Name Change
I have given all the Insurrectionist Units the SAS-10 as their handgun and the Officers the Comet. We have Innie weapons so let's utilise them. I also went and made it so the Innie assets can be called under the OPFOR tag as reinforcements making it easier for Zeus to create believable LZ's and RP's.